+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Removed `Notebook` `colab` and `jupyterlite` arguments, switching to always determining the interpreter automatically
 * Made `dill` a required dependency
 * Removed `variables` key of assignment config in favor of `generate.serialized_variables` in Otter Assign per [#628](https://github.com/ucbds-infra/otter-grader/issues/628)
+* Update Otter Assign to add cell metadata so that questions with no check cell have their tests run after the last solution cell per [#798](https://github.com/ucbds-infra/otter-grader/issues/798)
 
 **v5.6.0:**